requirements.txt
# FastAPI Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# LangChain & LLM
langchain==0.1.5
langchain-openai==0.0.5
langchain-community==0.0.16
# openai==1.10.0  # DeepSeek API兼容OpenAI SDK（已切换至智谱AI）

# 智谱AI SDK
zai-sdk>=0.2.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# WebSocket & Async
websockets==12.0
python-multipart==0.0.6
aiofiles==23.2.1

# Utilities
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Logging & Monitoring
loguru==0.7.2

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
